LayoutTests/platform/glib/http/tests/websocket/tests/hybi/reserved-opcodes-expected.txt

CONSOLE MESSAGE: WebSocket connection to 'ws://localhost:8880/websocket/tests/hybi/reserved-opcodes?opcode=3' failed: Received invalid WebSocket response from the server
CONSOLE MESSAGE: WebSocket connection to 'ws://localhost:8880/websocket/tests/hybi/reserved-opcodes?opcode=4' failed: Received invalid WebSocket response from the server
CONSOLE MESSAGE: WebSocket connection to 'ws://localhost:8880/websocket/tests/hybi/reserved-opcodes?opcode=5' failed: Received invalid WebSocket response from the server
CONSOLE MESSAGE: WebSocket connection to 'ws://localhost:8880/websocket/tests/hybi/reserved-opcodes?opcode=6' failed: Received invalid WebSocket response from the server
CONSOLE MESSAGE: WebSocket connection to 'ws://localhost:8880/websocket/tests/hybi/reserved-opcodes?opcode=7' failed: Received invalid WebSocket response from the server
CONSOLE MESSAGE: WebSocket connection to 'ws://localhost:8880/websocket/tests/hybi/reserved-opcodes?opcode=11' failed: Received invalid WebSocket response from the server
CONSOLE MESSAGE: WebSocket connection to 'ws://localhost:8880/websocket/tests/hybi/reserved-opcodes?opcode=12' failed: Received invalid WebSocket response from the server
CONSOLE MESSAGE: WebSocket connection to 'ws://localhost:8880/websocket/tests/hybi/reserved-opcodes?opcode=13' failed: Received invalid WebSocket response from the server
CONSOLE MESSAGE: WebSocket connection to 'ws://localhost:8880/websocket/tests/hybi/reserved-opcodes?opcode=14' failed: Received invalid WebSocket response from the server
CONSOLE MESSAGE: WebSocket connection to 'ws://localhost:8880/websocket/tests/hybi/reserved-opcodes?opcode=15' failed: Received invalid WebSocket response from the server
Tests whether WebSocket ignores frames with reserved opcode.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Opcode 3: Test started.
onopen() was called.
onclose() was called.
PASS closeEvent.wasClean is false
Opcode 4: Test started.
onopen() was called.
onclose() was called.
PASS closeEvent.wasClean is false
Opcode 5: Test started.
onopen() was called.
onclose() was called.
PASS closeEvent.wasClean is false
Opcode 6: Test started.
onopen() was called.
onclose() was called.
PASS closeEvent.wasClean is false
Opcode 7: Test started.
onopen() was called.
onclose() was called.
PASS closeEvent.wasClean is false
Opcode 11: Test started.
onopen() was called.
onclose() was called.
PASS closeEvent.wasClean is false
Opcode 12: Test started.
onopen() was called.
onclose() was called.
PASS closeEvent.wasClean is false
Opcode 13: Test started.
onopen() was called.
onclose() was called.
PASS closeEvent.wasClean is false
Opcode 14: Test started.
onopen() was called.
onclose() was called.
PASS closeEvent.wasClean is false
Opcode 15: Test started.
onopen() was called.
onclose() was called.
PASS closeEvent.wasClean is false
PASS successfullyParsed is true

TEST COMPLETE
